feat: Initial refactor, page creation, and bug fixes
This commit introduces several improvements and foundational elements:

1.  **Fix Balance Calculation:** Corrected the trainee balance calculation in `trainee.actions.ts` to use `purchase.totalPrice` instead of a non-existent `session.price`. The Prisma query was updated to include `purchases`.

2.  **Consolidate Prisma Client:** Removed the redundant `src/lib/db.ts` file. `src/lib/prisma.ts` is now the sole source for Prisma client instantiation.

3.  **Refine File Structure:**
    *   Colocated feature-specific components into their respective route directories within `src/app`. For example, `AuthForm.tsx` is now in `src/app/auth/_components/`, and `TraineeCard.tsx` is in `src/app/dashboard/trainees/_components/`.
    *   Moved `sidebar.tsx` and `topbar.tsx` (previously in `src/components/dashboard/`) to `src/app/dashboard/_components/`. Note: these components are not currently imported by the main dashboard layout.

4.  **Create Missing Dashboard Pages:**
    *   Added basic placeholder pages for:
        - `/dashboard/sessions`
        - `/dashboard/payments`
        - `/dashboard/schedule`
        - `/dashboard/profile`
        - `/dashboard/settings`
    *   These pages include basic titles and descriptive text in Hebrew.

Note: Linting and build checks could not be performed due to persistent errors in the execution environment.
