+++ src/app/dashboard/payments/page.tsx
@@ -1,0 +1,9 @@
+// src/app/dashboard/payments/page.tsx
+export default function PaymentsPage() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold">ניהול תשלומים</h1>
+      <p>כאן יוצגו ויתנהלו התשלומים והחיובים.</p>
+    </div>
+  );
+}
