+++ src/app/dashboard/profile/page.tsx
@@ -1,0 +1,9 @@
+// src/app/dashboard/profile/page.tsx
+export default function ProfilePage() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold">פרופיל אישי</h1>
+      <p>כאן יוצגו ויתעדכנו פרטי הפרופיל של המאמן.</p>
+    </div>
+  );
+}
