+++ src/app/dashboard/schedule/page.tsx
@@ -1,0 +1,9 @@
+// src/app/dashboard/schedule/page.tsx
+export default function SchedulePage() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold">לוח זמנים</h1>
+      <p>כאן יוצג ויתנהל לוח הזמנים של המאמן.</p>
+    </div>
+  );
+}
