+++ src/app/dashboard/sessions/page.tsx
@@ -1,0 +1,9 @@
+// src/app/dashboard/sessions/page.tsx
+export default function SessionsPage() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold">ניהול אימונים</h1>
+      <p>כאן יוצגו ויתנהלו האימונים של המתאמנים.</p>
+    </div>
+  );
+}
