+++ src/app/dashboard/settings/page.tsx
@@ -1,0 +1,9 @@
+// src/app/dashboard/settings/page.tsx
+export default function SettingsPage() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold">הגדרות</h1>
+      <p>כאן יוגדרו העדפות המערכת והחשבון.</p>
+    </div>
+  );
+}
