+++ src/app/dashboard/trainees/_components/TraineeCard.tsx
